s/Integer/Long/
git-svn-id: svn+ssh://svn.projects.openwide.fr/svnroot/owsi-tools/trunk/owsi-core@1181 d3474844-eb8e-4abc-b058-2b321fed648b

+++ owsi-core-components/owsi-core-component-jpa-security/src/main/java/fr/openwide/core/jpa/security/acl/service/AbstractCoreAclServiceImpl.java
@@ -98,7 +98,7 @@ public abstract class AbstractCoreAclServiceImpl extends JpaDaoSupport implement
 		try {
 			Class<?> clazz = Class.forName(objectIdentity.getType());
 			
-			if (GenericEntity.class.isAssignableFrom(clazz) && objectIdentity.getIdentifier() instanceof Integer) {
+			if (GenericEntity.class.isAssignableFrom(clazz) && objectIdentity.getIdentifier() instanceof Long) {
 				return (GenericEntity<Long, ?>) getEntity(clazz, objectIdentity.getIdentifier());
 			}
 		} catch (ClassNotFoundException e) {
